rename: restructure components files by usage
