+++ src/components/common/ScrapButton.tsx
