+++ src/components/index.ts
@@ -8,11 +8,15 @@ export { default as NoticeBanner } from './common/NoticeBanner';
 export { default as ExternalLinkChip } from './common/ExternalLinkChip';
 export { Tab, Tabs } from './common/Tab';
 export { default as SnackBar } from './common/SnackBar';
+export { default as ScrapButton } from './common/ScrapButton';
 
 //user
+//common
+export { ItemList, ItemGrid } from './user/common/Item';
+export { TimeChip, SoldOutChip, ExtendChip } from './user/common/Chip';
+//seller
 export { default as SellerProfileHeader } from './user/SellerProfileHeader';
-export { ItemList, ItemGrid } from './user/Item';
-export { TimeChip, SoldOutChip, ExtendChip } from './user/Chip';
+
 //seller
 //common
 export { SaveButton, DefaultButton, ModalButton } from './seller/common/Button';
@@ -25,6 +29,7 @@ export {
 } from './seller/common/DetailInput';
 export { default as CategoryChip } from './seller/common/CategoryChip';
 export { default as RadioInputSelector } from './seller/common/RadioInputSelector';
+
 //item
 export { CategoryMultiSelector } from './seller/item/registration/CategoryMultiSelector';
 export {
@@ -35,3 +40,6 @@ export { default as ExternalLinkBottomSheet } from './seller/item/administration
 export { default as AdminItemBottomSheet } from './seller/item/administration/AdminItemBottomSheet';
 export { default as RadioBottomSheet } from './seller/common/RadioBottomSheet';
 export { default as SellerMyItem } from './seller/item/administration/SellerMyItem';
+
+//my
+export { default as SellerMyProfileHeader } from './seller/my/SellerMyProfileHeader';
