+++ src/pages/index.ts
@@ -7,7 +7,7 @@ export { default as SellerProfile } from './user/sellerProfile/SellerProfile';
 
 // 셀러
 export { default as SellerMyProfile } from './seller/sellerMyPage/SellerMyProfile';
-export { default as SellerMyProfileHeader } from './seller/sellerMyPage/SellerMyProfileHeader';
+export { default as SellerMyProfileHeader } from '../components/seller/my/SellerMyProfileHeader';
 export { default as MySelectionTab } from './seller/sellerMyPage/MySelectionTab';
 export { default as MyStoredItemTab } from './seller/sellerMyPage/MyStoredItemTab';
 export { default as MyItemReviewTab } from './seller/sellerMyPage/MyItemReviewTab';
